+++ scripts/constants.js
@@ -25,6 +25,7 @@ export const GROUP = {
     complex_action: { id: 'complex_action', name: 'ARKHAM_HORROR.LABELS.Skills', type: 'system' },
     reactions: { id: 'reactions', name: 'ARKHAM_HORROR.KNACK_SHEET.RollKind.Reaction', type: 'system' },
     insight: { id: 'insight', name: 'ARKHAM_HORROR.LABELS.Insight', type: 'system' },
+    healing: { id: 'healing', name: 'ARKHAM_HORROR.HEALING.SectionTitle', type: 'system' },
     injury_trauma: { id: 'injury_trauma', name: 'ARKHAM_HORROR.LABELS.InjuriesTrauma', type: 'system' },
     weapons: { id: 'weapons', name: 'ARKHAM_HORROR.LABELS.Weapons', type: 'system' },
     spells: { id: 'spells', name: 'ARKHAM_HORROR.LABELS.Spells', type: 'system' },